Examples/education/course-student-portal/backend/courses/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import CourseViewSet, StudentViewSet, EnrollmentViewSet

router = DefaultRouter()
router.register('courses',     CourseViewSet,     basename='course')
router.register('students',    StudentViewSet,    basename='student')
router.register('enrollments', EnrollmentViewSet, basename='enrollment')

urlpatterns = [
    path('', include(router.urls)),
    path('health/', lambda req: __import__('django.http', fromlist=['JsonResponse']).JsonResponse({'status': 'ok'})),
]
